ONOS-7924 Fix unable to set role for OpenFlow switch

+++ protocols/openflow/ctl/src/main/java/org/onosproject/openflow/controller/impl/RoleManager.java
@@ -94,7 +94,7 @@ class RoleManager implements RoleHandler {
             roleToSend = OFNiciraControllerRole.ROLE_OTHER;
             log.debug("Sending Nx Role.SLAVE to switch {}.", sw);
         }
-        int xid = sw.getNextTransactionId();
+        long xid = sw.getNextTransactionId();
         OFExperimenter roleRequest = OFFactories.getFactory(OFVersion.OF_10)
                 .buildNiciraControllerRoleRequest()
                 .setXid(xid)
@@ -122,7 +122,7 @@ class RoleManager implements RoleHandler {
                     + " Should only be used for queries.", sw);
         }
 
-        int xid = sw.getNextTransactionId();
+        long xid = sw.getNextTransactionId();
         OFRoleRequest rrm = sw.factory()
                 .buildRoleRequest()
                 .setRole(roleToSend)
@@ -182,7 +182,7 @@ class RoleManager implements RoleHandler {
     @Override
     public synchronized RoleRecvStatus deliverRoleReply(RoleReplyInfo rri)
             throws SwitchStateException {
-        int xid = (int) rri.getXid();
+        long xid = rri.getXid();
         RoleState receivedRole = rri.getRole();
         RoleState expectedRole = pendingReplies.getIfPresent(xid);
